requirements.txt
flask==2.3.2
flask-cors==3.0.10
mysql-connector-python==8.1.0
python-dotenv==1.0.0
pyjwt==2.8.0
werkzeug==2.3.7